requirements.txt
SQLAlchemy==1.3.22
pandas==1.2.0
psycopg2==2.8.6
alembic==1.5.4
python-dotenv==0.15.0